@@ -1,15 +1,19 @@
 <template>
   <v-card>
-    <div v-if=shouldDisplayResponseCardTitle>
-      <v-card-title v-if="responseCard.title && responseCard.title.trim()" primary-title class="red lighten-5">
-        <span class="headline">{{responseCard.title}}</span>
+    <div v-if="shouldDisplayResponseCardTitle">
+      <v-card-title
+        v-if="responseCard.title && responseCard.title.trim()"
+        primary-title
+        class="red lighten-5"
+      >
+        <span class="headline">{{ responseCard.title }}</span>
       </v-card-title>
     </div>
     <v-card-text v-if="responseCard.subTitle">
-      <span>{{responseCard.subTitle}}</span>
+      <span>{{ responseCard.subTitle }}</span>
     </v-card-text>
     <v-card-text v-if="responseCard.subtitle">
-      <span>{{responseCard.subtitle}}</span>
+      <span>{{ responseCard.subtitle }}</span>
     </v-card-text>
     <v-card-media
       v-if="responseCard.imageUrl"
@@ -19,17 +23,17 @@
     ></v-card-media>
     <v-card-actions v-if="responseCard.buttons" class="button-row">
       <v-btn
-        v-for="(button) in responseCard.buttons"
+        v-for="button in responseCard.buttons"
         v-show="button.text && button.value"
         v-bind:key="button.id"
-        v-on:click.once.native="onButtonClick(button.value)"
+        v-on:click.once="onButtonClick(button.value)"
         v-bind:disabled="shouldDisableClickedResponseCardButtons"
         round
         default
         v-bind:color="button.text.toLowerCase() === 'more' ? '' : 'accent'"
         class="secondary--text"
       >
-        {{button.text}}
+        {{ button.text }}
       </v-btn>
     </v-card-actions>
     <v-card-actions v-if="responseCard.attachmentLinkUrl">
@@ -64,35 +68,35 @@ export default {
   props: ['response-card'],
   data() {
     return {
-      hasButtonBeenClicked: false,
-    };
+      hasButtonBeenClicked: false
+    }
   },
   computed: {
     shouldDisplayResponseCardTitle() {
-      return this.$store.state.config.ui.shouldDisplayResponseCardTitle;
+      return this.$store.state.config.ui.shouldDisplayResponseCardTitle
     },
     shouldDisableClickedResponseCardButtons() {
       return (
         this.$store.state.config.ui.shouldDisableClickedResponseCardButtons &&
         (this.hasButtonBeenClicked || this.getRCButtonsDisabled())
-      );
-    },
+      )
+    }
   },
-  inject: ['getRCButtonsDisabled','setRCButtonsDisabled'],
+  inject: ['getRCButtonsDisabled', 'setRCButtonsDisabled'],
   methods: {
     onButtonClick(value) {
-      this.hasButtonBeenClicked = true;
-      this.setRCButtonsDisabled();
-      const messageType = this.$store.state.config.ui.hideButtonMessageBubble ? 'button' : 'human';
+      this.hasButtonBeenClicked = true
+      this.setRCButtonsDisabled()
+      const messageType = this.$store.state.config.ui.hideButtonMessageBubble ? 'button' : 'human'
       const message = {
         type: messageType,
-        text: value,
-      };
+        text: value
+      }
 
-      this.$store.dispatch('postTextMessage', message);
-    },
-  },
-};
+      this.$store.dispatch('postTextMessage', message)
+    }
+  }
+}
 </script>
 
 <style scoped>
